Apply o11 MetricsManager
