+++ db/metrics/src/main/java/org/commonjava/indy/db/metered/MeasuringArtifactStoreQueryInterceptor.java
@@ -17,7 +17,7 @@ package org.commonjava.indy.db.metered;
 
 import org.commonjava.indy.data.ArtifactStoreQuery;
 import org.commonjava.indy.data.StoreDataManager;
-import org.commonjava.indy.subsys.metrics.IndyMetricsManager;
+import org.commonjava.o11yphant.metrics.DefaultMetricsManager;
 import org.commonjava.indy.model.core.ArtifactStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,7 +35,7 @@ public abstract class MeasuringArtifactStoreQueryInterceptor
     private StoreDataManager dataManager;
 
     @Inject
-    private IndyMetricsManager metricsManager;
+    private DefaultMetricsManager metricsManager;
 
     @Override
     public ArtifactStoreQuery<ArtifactStore> query()
