+++ subsys/jaxrs/src/main/java/org/commonjava/indy/bind/jaxrs/metrics/IndyHealthCheckDeploymentProvider.java
@@ -20,7 +20,7 @@ import io.undertow.servlet.Servlets;
 import io.undertow.servlet.api.DeploymentInfo;
 import io.undertow.servlet.api.ServletInfo;
 import org.commonjava.indy.bind.jaxrs.IndyDeploymentProvider;
-import org.commonjava.indy.subsys.metrics.IndyMetricsManager;
+import org.commonjava.o11yphant.metrics.DefaultMetricsManager;
 import org.commonjava.indy.subsys.metrics.conf.IndyMetricsConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,7 +39,7 @@ public class IndyHealthCheckDeploymentProvider
     private IndyMetricsConfig config;
 
     @Inject
-    private IndyMetricsManager metricsManager;
+    private DefaultMetricsManager metricsManager;
 
     @Override
     public DeploymentInfo getDeploymentInfo( String contextRoot, Application application )
